Refactoring Models
- Changed the way relationships are handled between score and organization
- fixed average scoring on company detail view
- separated models into individual files in crank/models
- broke and then fixed circular import problem
- added scores seeds
