+++ requirements.txt
@@ -1,3 +1,4 @@
-Django~=4.2.6
-fontawesomefree=6.4.2
-django-bootstrap5=5.3
+Django~=4.2
+fontawesomefree~=6.4
+django-bootstrap5~=23.3
+coverage~=7.3
